+++ packages/app/lib/storage/storage.ts
@@ -1,4 +1,4 @@
-import * as SecureStore from 'expo-secure-store'
+import AsyncStorage from '@react-native-async-storage/async-storage'
 
 import {
   API_TOKEN_STORAGE_KEY,
@@ -6,16 +6,30 @@ import {
   HEX_PUBKEY_STORAGE_KEY,
 } from './storage-constants'
 
-export function getItem(key) {
-  return SecureStore.getItemAsync(key)
+export async function getItem(key: string): Promise<string | null> {
+  try {
+    const value = await AsyncStorage.getItem(key)
+    return value
+  } catch (e) {
+    console.error('Error getting item from AsyncStorage:', e)
+    return null
+  }
 }
 
-export function setItem(key, value) {
-  return SecureStore.setItemAsync(key, value)
+export async function setItem(key: string, value: string): Promise<void> {
+  try {
+    await AsyncStorage.setItem(key, value)
+  } catch (e) {
+    console.error('Error setting item in AsyncStorage:', e)
+  }
 }
 
-export function removeItem(key) {
-  return SecureStore.deleteItemAsync(key)
+export async function removeItem(key: string): Promise<void> {
+  try {
+    await AsyncStorage.removeItem(key)
+  } catch (e) {
+    console.error('Error removing item from AsyncStorage:', e)
+  }
 }
 
 export async function getKeys() {
