+++ packages/app/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "@clerk/nextjs": "^4.6.15",
     "@my/ui": "*",
+    "@react-native-async-storage/async-storage": "~1.17.3",
     "@tamagui/animations-react-native": "^1.0.17",
     "@tamagui/colors": "^1.0.17",
     "@tamagui/font-inter": "^1.0.17",
